Updates packages adds support for React 15.5+
- Updates all of the dev dependencies
- Adds React and React-DOM to the dev dependencies because the react
  test utils are now located at react-dom/test-utils
- Adds budo to the dev dependencies
- Updates .eslintrc to handle changes from new versions of
    react-eslint
- Updates components to use prop-types package instead of
  React.PropTypes
- Adds PropTypes as a peer dependency
- Updates tests to use react-dom/test-utils
- Adds yarn.lock to be able to use yarn to install deps

One thing to note is that when performing an install there is a warning saying
that react-motion requires React v15.3.0, but looking at the package.json
in the react-motion project, the dev-dependencies include React v15.5.0 and
one of the dependencies declared is the prop-types package. This means the
peer dependencies need to be updated in react-motion.